twigane-models/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
pydantic==2.4.2
pydantic-settings==2.0.3
transformers==4.35.0
torch==2.1.0
opencv-python==4.8.1.78
python-dotenv==1.0.0
motor==3.3.1
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
tensorflow==2.14.0
tensorflow-hub==0.14.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pandas==2.1.2
numpy>=1.23.5,<2.0.0
pillow==10.1.0
scikit-learn==1.3.2
aiofiles==23.2.1
websockets==12.0
pymongo==4.6.0
jinja2==3.1.2
email-validator==2.1.0.post1
python-socketio==5.10.0
redis==5.0.1
celery==5.3.4
boto3==1.28.64
requests==2.31.0
TTS==0.17.6